fix type mismatch during memcpy

+++ src/hdtv/rootext/mfile-root/mfile/src/txt_getput.c
@@ -42,7 +42,7 @@ int32_t txt_get(MFILE *mat, void *buffer, int32_t level, int32_t line, int32_t c
 
   int32_t idx = (((uint32_t)level * mat->lines) + (uint32_t)line) * mat->columns + (uint32_t)col;
 
-  memcpy((int32_t *)buffer, dblp + idx, (uint32_t)num * sizeof(double));
+  memcpy((double *)buffer, dblp + idx, (uint32_t)num * sizeof(double));
 
   return num;
 }
@@ -54,7 +54,7 @@ int32_t txt_put(MFILE *mat, void *buffer, int32_t level, int32_t line, int32_t c
 
   int32_t idx = (((uint32_t)level * mat->lines) + (uint32_t)line) * mat->columns + (uint32_t)col;
 
-  memcpy(dblp + idx, (int32_t *)buffer, (uint32_t)num * sizeof(double));
+  memcpy(dblp + idx, (double *)buffer, (uint32_t)num * sizeof(double));
 
   return num;
 }
